==> code/1 Load_data_and_EDA/4_corpus_statistics.ipynb <==
import os
import json
import pandas as pd
annotations_path = 'C:/Users/pdkud/Documents/Learning/Магистратура МФТИ/RSL_corpus/annotations'

annotations = [f for f in os.listdir(annotations_path) if f != 'description.csv' and 'bad' not in f]

translations_list = list()
for filename in annotations:
    translations_dict = dict()
    with open(annotations_path+'/'+filename, 'r') as f:
        data = json.load(f)
        for r in data:
            translations_dict[(r[1], r[2])] = r[0]
    translations_list.append([filename, translations_dict])

translations_cnt = sum([len(x[1]) for x in translations_list])

print(f"Общее кол-во слов и файлов, выделенных на основе аннотации: {translations_cnt}")

translations = pd.read_excel('C:/Users/pdkud/Documents/Learning/Магистратура МФТИ/RSL_corpus/translations.xlsx')
translations['LENGTH'] = translations['TRANSLATION'].str.split(' ').str.len()

translations['CLEANED'] = translations.TRANSLATION.str.lower()\
    .str.replace("[.?!,:.\"–=#\^]", "", regex=True)\
    .str.replace("clf", "").str.replace("indx", "").str.replace("1ps", "")\
    .str.replace("du", "").str.replace("fut", "").str.replace("imp", "")\
    .str.replace("loc", "").str.replace("ord", "").str.replace("past", "")\
    .str.replace("pi", "").str.replace("pres", "").str.replace("prtcl", "")

# Уникальные глоссы
len(set(translations[translations['LENGTH']==1.0].CLEANED))

# Уникальные фразы
len(set(translations[translations['LENGTH'] > 1.0].CLEANED))

translations[translations['LENGTH']==1.0].CLEANED

description = pd.read_csv(
    'C:/Users/pdkud/Documents/Learning/Магистратура МФТИ/RSL_corpus/annotations/description.csv',
    encoding = 'latin-1',
    delimiter=';')

len(set(description.respondent_id))

description.item_city.unique()
