==> code/1 Load_data_and_EDA/4_slovo_statistics.ipynb <==
# EDA: Slovo

annotations_path = 'C:/Users/pdkud/Documents/Learning/Магистратура МФТИ/slovo/slovo_annotations/SLOVO_DATAFRAME.tsv'

import os
import pandas as pd

slovo_df = pd.read_csv(annotations_path, sep='\t')

slovo_df.head()

slovo_df['frames_cnt'] = slovo_df['end']-slovo_df['begin']+1
slovo_df.head()

slovo_df.groupby(['height', 'width']).agg({'attachment_id' : ['count'], 'frames_cnt':['sum']}).sort_values(['height', 'width'])

slovo_df.agg({'attachment_id' : ['count'], 'frames_cnt':['sum']})

len(slovo_df.text.unique())

slovo_df['CLEANED'] = slovo_df.text.str.lower()\
    .str.replace("[.?!,:.\"–=#\^]", "", regex=True)

slovo_df['LENGTH'] = slovo_df['CLEANED'].str.split(' ').str.len()

# Уникальные фразы
len(set(slovo_df[slovo_df['LENGTH'] > 1.0].CLEANED))
